requirements.txt
django
gunicorn
django-heroku
pdfminer2
pandas
requests
django-background-tasks
